convert shader
