+++ Engine/src/component/Component.cpp
@@ -67,16 +67,6 @@ void ObjectComponent::attachToEntity(std::shared_ptr<Component> c, Entity entity
 	}
 }
 
-void ShaderComponent::attachToEntity(std::shared_ptr<Component> c, Entity entityHandler, Scene& scene)
-{
-	(void)scene;
-	if (auto sc = std::dynamic_pointer_cast<ShaderComponent>(c))
-	{
-		auto& shader = entityHandler.addComponent<ShaderComponent>(*sc);
-		shader.update();
-	}
-}
-
 void InstanceBatch::attachToEntity(std::shared_ptr<Component> c, Entity entityHandler, Scene& scene)
 {
         (void)scene;
@@ -166,120 +156,6 @@ void InstanceBatch::build()
 	glVertexAttribDivisor(7, 1);
 	glVertexAttribDivisor(8, 1);
 	glVertexAttribDivisor(9, 1);
-}
-
-ShaderComponent::ShaderComponent()
-{
-	renderViewProjection = std::make_shared<RenderView>(Viewport{ 0, 0, 1920, 1080 }, Entity::EmptyEntity);
-};
-
-void ShaderComponent::setProjectionTexture(Resource<Texture> texture)
-{
-	renderViewProjection->bind();
-	renderViewProjection->setTexture(texture);
-	projectionTexture = texture;
-	
-}
-
-#include <regex>
-
-void ShaderComponent::parseUniforms(const std::string& sourceCode)
-{
-	m_uniformProperties.clear();
-	customTextures.clear();
-
-	std::regex uniformRegex(R"(uniform\s+(\w+)\s+(\w+)\s*;)");
-	std::smatch match;
-	std::string::const_iterator searchStart(sourceCode.cbegin());
-
-	auto& uniformProperties = m_uniformProperties;
-
-	while (std::regex_search(searchStart, sourceCode.cend(), match, uniformRegex)) {
-		std::string type = match[1].str();
-		std::string name = match[2].str();
-
-		if (type == "float") {
-			uniformProperties[name] = 0.0f;
-		}
-		else if (type == "vec2") {
-			uniformProperties[name] = glm::vec2(0.0f);
-		}
-		else if (type == "vec3") {
-			uniformProperties[name] = glm::vec3(0.0f);
-		}
-		else if (type == "vec4") {
-			uniformProperties[name] = glm::vec4(0.0f);
-		}
-		else if (type == "int") {
-			uniformProperties[name] = 0;
-		}
-		else if (type == "uint") {
-			uniformProperties[name] = 0u;
-		}
-		else if (type == "mat3") {
-			uniformProperties[name] = glm::mat3(1.0f);
-		}
-		else if (type == "mat4") {
-			uniformProperties[name] = glm::mat4(1.0f);
-		}
-		else if (type == "sampler2D") {
-			customTextures[name] = Engine::get()->getCommonTextures()->getTexture(CommonTextures::TextureType::WHITE_1X1);
-		}
-
-		searchStart = match.suffix().first;
-	}
-}
-
-void ShaderComponent::setShader(Resource<Shader> shader)
-{
-	m_customShader = shader;
-
-	shaderOverride = shader->getShaderOverride();
-
-	const std::string& sourceCode = shader->getSourceCode();
-
-	parseUniforms(sourceCode);
-
-	isValid = true;
-}
-
-void ShaderComponent::update()
-{
-	auto oldUniforms = m_uniformProperties;
-	auto oldTextures = customTextures;
-
-	parseUniforms(m_customShader->getSourceCode());
-
-
-	auto& newTextures = customTextures;
-	for (const auto [name, texture] : oldTextures)
-	{
-		auto iter = newTextures.find(name);
-		if (iter != newTextures.end())
-		{
-			iter->second = texture;
-		}
-	}
-
-	auto& newUniforms = m_uniformProperties;
-	for (const auto [name, value] : oldUniforms)
-	{
-		auto iter = newUniforms.find(name);
-		if (iter != newUniforms.end())
-		{
-			iter->second = value;
-		}
-	}
-
-	for (const auto& [name, value] : m_uniformProperties)
-	{
-		m_customShader.get()->setUniformValue(name, value);
-	}
-
-	if (!projectionTexture.isEmpty())
-	{
-		setProjectionTexture(projectionTexture);
-	}
 }
 
 SkyboxComponent::SkyboxComponent(Resource<Texture> skyboxImage)
