+++ Engine/src/component/Component.h
@@ -176,62 +176,6 @@ struct EngineAPI ObjectComponent : public Component
     
 };
 
-
-struct EngineAPI ShaderComponent : public Component
-{
-	enum ProjectionType : int
-	{
-		DefaultProjection = 0,
-		Texture2D = 1
-	};
-
-	ShaderComponent();
-	
-	ShaderComponent(Shader* vertexShader, Shader* fragmentShader);
-
-	void addTexture(const std::string& name, Resource<Texture> texture)
-	{
-		customTextures[name] = texture;
-	}
-
-	void setProjectionTexture(Resource<Texture> texture);
-
-	void update();
-
-	void parseUniforms(const std::string& sourceCode);
-
-	void setShader(Resource<Shader> shader);
-
-	template <class Archive>
-	void serialize(Archive& archive) {
-			SERIALIZED_MEMBER(m_customShader); 
-			SERIALIZED_MEMBER(shaderOverride); 
-			SERIALIZED_MEMBER(customTextures); 
-			SERIALIZED_MEMBER(projection); 
-			SERIALIZED_MEMBER(projectionTexture);
-			SERIALIZED_MEMBER(m_shaderFilePath); 
-			SERIALIZED_MEMBER(isValid);
-	}
-
-	// This will only be used by forward renderer, ignored by deffered
-	Resource<Shader> m_customShader;
-
-	std::map<std::string, Resource<Texture>> customTextures;
-
-	ProjectionType projection = ProjectionType::DefaultProjection;
-	Resource<Texture> projectionTexture;
-	std::shared_ptr<RenderView> renderViewProjection;
-
-	std::unordered_map<std::string, Value> m_uniformProperties;
-
-	std::string m_shaderFilePath;
-	ShaderOverride shaderOverride;
-
-        bool isValid = false;
-        static void attachToEntity(std::shared_ptr<Component>, Entity, Scene&);
-};
-
-
 struct EngineAPI InstanceBatch : public Component
 {
 public:
