+++ Engine/src/render/DeferredRenderer.cpp
@@ -23,6 +23,7 @@
 #include "utils/DebugHelper.h"
 #include "component/MeshComponent.h"
 #include "component/MaterialComponent.h"
+#include "component/ShaderComponent.h"
 
 static float lerp(float a, float b, float t)
 {
